Contract tests vs functional tests
Contract tests can be seen as a subset of functional testing, limited to the behavior and functionality that the consumer needs. To state that consumer testing is not about behavior and functionality is confusing. The URL being used is "contract_tests_vs_functional_tests" instead of "contract_tests_not_functional_tests" which reflects this subtle difference.